fix: fix issue with crypto-js and pin dependencies
Fixed a bug where the registration system was no longer able to communicate with the listener. This was caused by a caret on `"crypto-js": "^4.1.1"` which caused version 4.2.0 to be installed, a version that [changed the default hashing algorithm](https://github.com/brix/crypto-js#420). This led to a mismatch in the hashing algorithms used for encryption/decryption in the site vs the listener.

Pinned the `crypto-js` dependency version to fix this issue, and pinned all other dependencies in this project to prevent similar issues.

See merge request opalmedapps/opal-registration#73