Add Docker files to be able to run registration web page in a container
Uses a multistage build where the dependencies are in their separate stage.
The only disadvantage is that npm and bower are not available at runtime.

If this is desired, the build should produce different results for dev and prod environments.


(cherry picked from commit cbacb5fd8fbc03972d3400e7d0543b878efaf344)

Co-authored-by: Matthias Schoettle <matthias.schoettle@muhc.mcgill.ca>